+++ tests/OTPFacadeTest.php
@@ -9,7 +9,7 @@ use PHPUnit\Framework\Attributes\Test;
 
 class OTPFacadeTest extends TestCase
 {
-    protected const RECIPIENT = '09389599530';
+    protected const RECIPIENT = '5555555555';
 
     #[Test]
     public function it_can_send_and_validate_a_token_via_the_facade(): void